done with refactoring the code and used a new hook useMemo
