+++ .eslintrc.json
@@ -3,5 +3,8 @@
   "env": {
     "es6": true,
     "browser": true
+  },
+  "rules": {
+    "react-hooks/exhaustive-deps": "error"
   }
 }
