+++ src/helpers.js
@@ -1,0 +1,9 @@
+import { Config, Direction } from "./constants";
+
+export const getRandomCell = () => ({
+  x: Math.floor(Math.random() * Config.width),
+  y: Math.floor(Math.random() * Config.width),
+  createdAt: Date.now(),
+});
+
+export const getInitialDirection = () => Direction.Right;
